Bumped package version to 1.2.0-alpha-1 Added Babel support to the script compiler plugin, as per pbastowski:angular-babel. The plugin first compiles the ES6 code with Babel and then imediately runs ng-annotate on the compiled code.
