+++ plugin/ng-script-compiler.js
@@ -1,8 +1,8 @@
 var to5 = Npm.require('babel-core');
 var fs = Npm.require('fs');
 var path = Npm.require('path');
-var ngAnnotate = Npm.require('ng-annotate');
 
+var ngAnnotate = Npm.require('ng-annotate');
 
 Object.merge = function (destination, source) {
   for (var property in source)
@@ -10,78 +10,71 @@ Object.merge = function (destination, source) {
   return destination;
 }
 
-var defaultConfig = {
+var config = {
   // print loaded config
   "debug":      false,
   // print active file extensions
   "verbose":    true,
-  // babel managed extensions
-  "extensions": ['es6.js', 'js', 'es6', 'jsx'],
   // experimental ES7 support
   "stage":      0,
   // module format to use
   "modules":    'common'
 }
 
-var handler = function (compileStep, isLiterate) {
-  var source = compileStep.read().toString('utf8');
-  //var outputFile = compileStep.inputPath + ".js";
-  var outputFile = compileStep.inputPath;
+var processFiles = function(files) {
+  files.forEach(processFile);
+};
+
+var processFile = function(file) {
+
+  var source = file.getContentsAsString();
+  var outputFile = file.getPathInPackage();
   var to5output = "";
+
+  console.log('Babel compiling: ', outputFile);
 
   try {
     to5output = to5.transform(source, {
+      // The blacklisting of "userStrict" is required to support
+      // Meteor's file level declarations that Meteor can export
+      // from packages.
       blacklist: ["useStrict"],
       sourceMap: true,
       stage:     config.stage,
-      filename:  compileStep.pathForSourceMap,
+      filename:  file.getDisplayPath(),
       modules:   config.modules
     });
   } catch (e) {
     console.log(e); // Show the nicely styled babel error
-    return compileStep.error({
-      message:    'Babel transform error',
-      sourcePath: compileStep.inputPath,
-      line:       e.loc.line,
-      column:     e.loc.column
+    return file.error({
+      message: 'Babel transform error',
+      line:    e.loc.line,
+      column:  e.loc.column
     });
   }
 
-  var ret = ngAnnotate(to5output.code, {
+  //var annotated = { src: to5output.code };
+
+  //to5output = source;
+
+  var annotated = ngAnnotate(to5output.code, {
     add: true
   });
 
-  if (ret.errors) {
-    throw new Meteor.Error(ret.errors.join(': '));
+  if (annotated.errors) {
+    throw new Error(annotated.errors.join(': \n\n'));
   }
 
-  compileStep.addJavaScript({
-    path:       outputFile,
-    sourcePath: compileStep.inputPath,
-    data:       ret.src,
-    sourceMap:  JSON.stringify(to5output.map)
+  file.addJavaScript({
+    data: annotated.src,
+    path: outputFile
   });
 };
 
-// initialization once at `meteor` exec
-var appdir = process.env.PWD || process.cwd();
-var filename = path.join(appdir, 'babel.json');
-var userConfig = {};
+Plugin.registerCompiler({
+  extensions: ['js'],
+  filenames: []
 
-if (fs.existsSync(filename)) {
-  userConfig = JSON.parse(fs.readFileSync(filename, {encoding: 'utf8'}));
-}
-
-var config = Object.merge(defaultConfig, userConfig);
-
-config.extensions.forEach(function (ext) {
-  Plugin.registerSourceHandler(ext, handler);
+}, function() {
+  return { processFilesForTarget: processFiles };
 });
-
-if (config.verbose)
-  console.log("Babel active on file extensions: " + config.extensions.join(', '));
-
-if (config.debug) {
-  console.log("\nBabel config:");
-  console.log(config);
-}
